+++ app/user/main/reducer.js
@@ -1,13 +1,13 @@
 import { combineReducers } from 'redux';
 
 /* 引入reducers */
-import workReducers from './reducers/workReducers';
+import workReducer from './reducers/workReducer';
 
-const mainRedux = {
+const mainReducers = {
   mount: 'main', // 挂载点
   reducer: combineReducers({
-    work: workReducers,
+    work: workReducer,
   }), // reducers
 };
 
-export default mainRedux;
+export default mainReducers;
